refactor: separated Socials and Copyright components from Footer for reusability
