+++ frontend/src/components/footer/Footer.jsx
@@ -1,12 +1,7 @@
-import {
-  FaLinkedin,
-  FaGithub,
-  FaXTwitter,
-  FaDiscord,
-  FaYoutube,
-} from 'react-icons/fa6';
 import { motion } from 'framer-motion';
 import { NavLink } from 'react-router-dom';
+import Socials from '../socials/Socials';
+import Copyright from '../copyright/Copyright';
 
 export default function Footer() {
   const navLinks = [
@@ -25,34 +20,6 @@ export default function Footer() {
     { title: 'Contact', path: '/contact' },
     { title: 'Terms & Conditions', path: '/terms-and-conditions' },
     { title: 'Privacy Policy', path: '/privacy-policy' },
-  ];
-
-  const socials = [
-    {
-      title: 'X / Twitter',
-      url: 'https://x.com/nagcas/',
-      icon: <FaXTwitter className='text-2xl' />,
-    },
-    {
-      title: 'LinkedIn',
-      url: 'https://www.linkedin.com/in/gianluca-chiaravalloti-5694081a2/',
-      icon: <FaLinkedin className='text-2xl' />,
-    },
-    {
-      title: 'GitHub',
-      url: 'https://github.com/nagcas/',
-      icon: <FaGithub className='text-2xl' />,
-    },
-    {
-      title: 'Discord',
-      url: 'https://discord.gg/RDBp8KJB',
-      icon: <FaDiscord className='text-2xl' />,
-    },
-    {
-      title: 'YouTube',
-      url: 'https://www.youtube.com/@TerraQuakeAPI',
-      icon: <FaYoutube className='text-2xl' />,
-    },
   ];
 
   return (
@@ -132,49 +99,11 @@ export default function Footer() {
         </div>
 
         {/* Social / Sponsor */}
-        <div>
-          <h3 className='text-white font-semibold mb-6 text-lg after:content-[""] after:block after:max-w-1/3 after:h-[2px] after:bg-violet-500 after:mt-2 after:mx-auto'>
-            Community & Socials
-          </h3>
-          <div className='flex flex-wrap justify-center gap-4'>
-            {socials.map((item) => (
-              <motion.a
-                key={item.title}
-                href={item.url}
-                target='_blank'
-                whileHover={{ scale: 1.2, y: -3 }}
-                transition={{ type: 'spring', stiffness: 200, damping: 10 }}
-                className='hover:text-violet-400 transition duration-300 text-slate-400'
-                aria-label={`Visit the TerraQuake API ${item.title} profile`}
-              >
-                {item.icon}
-              </motion.a>
-            ))}
-          </div>
-        </div>
+        <Socials />
       </div>
 
       {/* Copyright + Extended License */}
-      <div className='relative border-t border-white/10 mt-10 pt-6 px-6 text-center text-sm text-slate-400 space-y-4'>
-        <p>
-          &copy; {new Date().getFullYear()} TerraQuake API · All rights reserved
-        </p>
-        <p className='max-w-3xl mx-auto leading-relaxed px-4'>
-          TerraQuake API is free software: you can redistribute it and/or modify
-          it under the terms of the{' '}
-          <a
-            href='https://www.gnu.org/licenses/agpl-3.0.html'
-            target='_blank'
-            rel='noopener noreferrer'
-            className='hover:text-violet-400 transition duration-200'
-            aria-label='Visit the TerraQuake API licence'
-          >
-            GNU Affero General Public License
-          </a>{' '}
-          as published by the Free Software Foundation, either version 3 of the
-          License, or (at your option) any later version.
-        </p>
-      </div>
+      <Copyright />
 
       <div className='absolute bottom-0 left-0 right-0 h-[2px] bg-gradient-to-r from-violet-500 via-fuchsia-500 to-blue-500 opacity-70 blur-sm'></div>
     </footer>
